+++ home.admin/config.scripts/blitz.setpassword.sh
@@ -3,10 +3,11 @@
 # command info
 if [ "$1" = "-h" ] || [ "$1" = "-help" ]; then
  echo "small config script to set a passwords A,B,C & D"
- echo "blitz.setpassword.sh [?a|b|c|d] [?newpassword] "
+ echo "blitz.setpassword.sh a [?newpassword] "
+echo "blitz.setpassword.sh b [?newpassword] "
+ echo "blitz.setpassword.sh c [?oldpassword] [?newpassword] "
  echo "or just as a password enter dialog (result as file)"
  echo "blitz.setpassword.sh [x] [text] [result-file] [?empty-allowed]"
- echo "exits on 0 = needs reboot"
  exit 1
 fi
 
@@ -29,26 +30,18 @@ if [ ${#chain} -eq 0 ]; then
   chain="main"
 fi
 
-# 1. parameter [?a|b|c|d]
+# 1. parameter [?a|b|c]
 abcd=$1
 
-# 2. parameter [?newpassword]
-newPassword=$2
-
-# 4. parameter [?newpassword]
-emptyAllowed=0
-if [ "$4" == "empty-allowed" ]; then
-  emptyAllowed=1
-fi
-
 # run interactive if no further parameters
+reboot=0;
 OPTIONS=()
 if [ ${#abcd} -eq 0 ]; then
+    reboot=1;
     emptyAllowed=1
     OPTIONS+=(A "Master User Password / SSH")
     OPTIONS+=(B "RPC Password (blockchain/lnd)")
     OPTIONS+=(C "LND Wallet Password")
-    OPTIONS+=(D "LND Seed Password")
     CHOICE=$(dialog --clear \
                 --backtitle "RaspiBlitz" \
                 --title "Set Password" \
@@ -82,6 +75,8 @@ echo ""
 ############################
 # PASSWORD A
 if [ "${abcd}" = "a" ]; then
+
+  newPassword=$2
 
   # if no password given by parameter - ask by dialog
   if [ ${#newPassword} -eq 0 ]; then
@@ -152,20 +147,21 @@ if [ "${abcd}" = "a" ]; then
 
   echo ""
   echo "OK - password A changed for user pi, root, admin & bitcoin"
-  exit 0
 
 ############################
 # PASSWORD B
 elif [ "${abcd}" = "b" ]; then
+
+  newPassword=$2
 
   # if no password given by parameter - ask by dialog
   if [ ${#newPassword} -eq 0 ]; then
     clear
 
     # ask user for new password B (first time)
-    password1=$(whiptail --passwordbox "\nPlease enter your RPC Password B:\n(min 8chars, 1word, chars+number, no specials)" 10 52 "" --title "Password B" --backtitle "RaspiBlitz - Setup" 3>&1 1>&2 2>&3)
+    password1=$(whiptail --passwordbox "\nPlease enter your new Password B:\n(min 8chars, 1word, chars+number, no specials)" 10 52 "" --title "Password B" --backtitle "RaspiBlitz - Setup" 3>&1 1>&2 2>&3)
     if [ $? -eq 1 ]; then
-      if [ ${emptyAllowed} -eq 0 ]; then
+      if [ "${emptyAllowed}" == "0" ]; then
         echo "CANCEL not possible"
         sleep 2
       else
@@ -176,7 +172,7 @@ elif [ "${abcd}" = "b" ]; then
     # ask user for new password B (second time)
     password2=$(whiptail --passwordbox "\nRe-Enter Password B:\n" 10 52 "" --title "Password B" --backtitle "RaspiBlitz - Setup" 3>&1 1>&2 2>&3)
     if [ $? -eq 1 ]; then
-      if [ ${emptyAllowed} -eq 0 ]; then
+      if [ "${emptyAllowed}" == "0" ]; then
         echo "CANCEL not possible"
         sleep 2
       else
@@ -287,60 +283,119 @@ EOF
   # ThunderHub
   if [ "${thunderhub}" == "on" ]; then
     echo "# changing the password for ThunderHub"
-    sed -i "s/^masterPassword:.*/masterPassword: '${newPassword}'/g" /mnt/hdd/app-data/thunderhub/thubConfig.yaml
+    sudo sed -i "s/^masterPassword:.*/masterPassword: '${newPassword}'/g" /mnt/hdd/app-data/thunderhub/thubConfig.yaml
+  fi
+
+  # LIT
+  if [ "${lit}" == "on" ]; then
+    echo "# changing the password for LIT"
+    sudo sed -i "s/^uipassword=.*/uipassword=${newPassword}/g" /mnt/hdd/app-data/.lit/lit.conf
+    sudo sed -i "s/^faraday.bitcoin.password=.*/faraday.bitcoin.password=${newPassword}/g" /mnt/hdd/app-data/.lit/lit.conf
   fi
 
   echo "# OK -> RPC Password B changed"
   echo "# Reboot is needed"
-  exit 0
 
 ############################
 # PASSWORD C
 elif [ "${abcd}" = "c" ]; then
 
-  if [ ${#newPassword} -gt 0 ]; then
-    echo "New password C cannot be set thru paramter .. will start interactive password setting."
-    echo "PRESS ENTER to continue"
-    read key
+  oldPassword=$2
+  newPassword=$3
+
+  if [ "${oldPassword}" == "" ]; then
+    # ask user for old password c
+    clear
+    oldPassword=$(whiptail --passwordbox "\nEnter old Password C:\n" 10 52 "" --title "Old Password C" --backtitle "RaspiBlitz - Passwords" 3>&1 1>&2 2>&3)
+    if [ $? -eq 1 ] || [ "${oldPassword}" == "" ]; then
+      sudo /home/admin/config.scripts/blitz.setpassword.sh c
+    fi
+    echo "OK ... processing"
   fi
 
-  clear
-  echo ""
-  echo "****************************************************************************"
-  echo "Change LND Wallet Password --> lncli --chain=${network} --network=${chain}net changepassword"
-  echo "****************************************************************************"
-  echo "This is your Password C on the RaspiBlitz to unlock your LND wallet."
-  echo "If you had Auto-Unlock active - you need to re-activate after this."
-  echo "****************************************************************************"
+  if [ "${newPassword}" == "" ]; then
+    clear
+
+    # ask user for new password c
+    newPassword=$(whiptail --passwordbox "\nEnter new Password C:\n" 10 52 "" --title "New Password C" --backtitle "RaspiBlitz - Passwords" 3>&1 1>&2 2>&3)
+    if [ $? -eq 1 ] || [ "${newPassword}" == "" ]; then
+      sudo /home/admin/config.scripts/blitz.setpassword.sh c ${oldPassword}
+      exit 0
+    fi
+    # check new password does not contain bad characters
+    clearedResult=$(echo "${newPassword}" | tr -dc '[:alnum:]-.' | tr -d ' ')
+    if [ ${#clearedResult} != ${#newPassword} ] || [ ${#clearedResult} -eq 0 ]; then
+      dialog --backtitle "RaspiBlitz - Setup" --msgbox "FAIL -> Contains bad characters (spaces, special chars)" 6 52
+      sudo /home/admin/config.scripts/blitz.setpassword.sh c ${oldPassword}
+      exit 0
+    fi
+    # check new password longer than 8
+    if [ ${#newPassword} -lt 8 ]; then
+      dialog --backtitle "RaspiBlitz - Setup" --msgbox "FAIL -> Password length under 8" 6 52
+      sudo /home/admin/config.scripts/blitz.setpassword.sh c ${oldPassword}
+      exit 1
+    fi
+
+    # ask user to retype new password c
+    newPassword2=$(whiptail --passwordbox "\nEnter again new Password C:\n" 10 52 "" --title "New Password C (repeat)" --backtitle "RaspiBlitz - Passwords" 3>&1 1>&2 2>&3)
+    if [ $? -eq 1 ] || [ "${newPassword}" == "" ]; then
+      sudo /home/admin/config.scripts/blitz.setpassword.sh c ${oldPassword}
+      exit 0
+    fi
+    echo "OK ... processing"
+    # check if passwords match
+    if [ "${newPassword}" != "${newPassword2}" ]; then
+      dialog --backtitle "RaspiBlitz - Setup" --msgbox "FAIL -> Passwords dont Match" 6 52
+      sudo /home/admin/config.scripts/blitz.setpassword.sh c ${oldPassword}
+      exit 0
+    fi
+    echo "OK ... processing"
+  fi
+
+  #echo "oldPassword: ${oldPassword}"
+  #echo "newPassword: ${newPassword}"
+
+  echo "# Make sure Auto-Unlocks off"
+  sudo /home/admin/config.scripts/lnd.autounlock.sh off
 
   echo "LND needs to be restarted to lock wallet first .. (please wait)"
   sudo systemctl restart lnd
-  sleep 6
+  sleep 2
 
+  err=""
+  source <(sudo /home/admin/config.scripts/lnd.initwallet.py change-password $oldPassword $newPassword)
+  if [ "${err}" != "" ]; then
+    dialog --backtitle "RaspiBlitz - Setup" --msgbox "FAIL -> Was not able to change password\n\n${err}\n${errMore}" 10 52
+    clear
+    echo "# FAIL: Was not able to change password"
+    exit 1
+  fi
+
+  # old manual way
+  # clear
+  # echo ""
+  # echo "****************************************************************************"
+  # echo "Change LND Wallet Password --> lncli --chain=${network} --network=${chain}net changepassword"
+  # echo "****************************************************************************"
+  # echo "This is your Password C on the RaspiBlitz to unlock your LND wallet."
+  # echo "If you had Auto-Unlock active - you need to re-activate after this."
+  # echo "****************************************************************************"
+  # sleep 6
   # let LND-CLI handle the password change
-  sudo -u bitcoin lncli --chain=${network} --network=${chain}net changepassword
-
-  # deactivate AUTO-UNLOCK if activated
-  echo ""
-  echo "# Make sure Auto-Unlocks off"
-  sudo /home/admin/config.scripts/lnd.autounlock.sh off
+  # sudo -u bitcoin lncli --chain=${network} --network=${chain}net changepassword
 
   # final user output
   echo ""
   echo "OK"
-  exit 0
-
-############################
-# PASSWORD D
-elif [ "${abcd}" = "d" ]; then
-
-  echo "#### NOTICE ####"
-  echo "Sorry - the password D cannot be changed. Its the password you set on creating your wallet to protect your seed (the list of words)."
-  exit 1
 
 ############################
 # PASSWORD X
 elif [ "${abcd}" = "x" ]; then
+
+    emptyAllowed=0
+    if [ "$4" == "empty-allowed" ]; then
+      emptyAllowed=1
+    fi
 
     # second parameter is the flexible text
     text=$2
@@ -396,4 +451,12 @@ elif [ "${abcd}" = "x" ]; then
 else
   echo "FAIL: there is no password '${abcd}' (reminder: use lower case)"
   exit 1
+fi
+
+# when started with menu ... reboot when done
+if [ "${reboot}" == "1" ]; then
+  echo "Now rebooting to activate changes ..."
+  sudo /home/admin/XXshutdown.sh reboot
+else
+  echo "..."
 fi