+++ home.admin/config.scripts/bonus.sphinxrelay.sh
@@ -26,7 +26,7 @@ if [ "$1" = "menu" ]; then
     whiptail --title " Warning " \
     --yes-button "Back" \
     --no-button "Continue Anyway" \
-    --yesno "Your SPHINX SERVER may have problems (retry if just restarted).\n\nCheck if locally responding: http://${localIP}:${httpPort}/app\n(You should see 'INDEX' in your browser)\n\nCheck if service is reachable over Tor:\n${toraddress}/app\n\nAlso check logs with 'debug' on terminal." 16 72
+    --yesno "Your SPHINX SERVER may have problems (retry if just restarted).\n\nCheck if locally responding: http://${localIP}:${httpPort}/app\n(You should see 'INDEX' in your browser)\n\nCheck if service is reachable over Tor:\n${toraddress}/app\n\nIf you use IP2TOR may cancel & renew subscription & LetsEncrypt.\nAlso check logs with 'debug' on terminal." 17 72
     if [ "$?" != "1" ]; then
       exit 0
 	  fi
@@ -34,7 +34,7 @@ if [ "$1" = "menu" ]; then
     whiptail --title " Warning " \
     --yes-button "Back" \
     --no-button "Continue Anyway" \
-    --yesno "Your HTTPS connection over IP2TOR as has problems:\n${ip2torWarn}\n\nCheck if service is reachable over Tor:\n${toraddress}/app\n\nMaybe cancel the IP2Tor & LetsEncrypt subscription & setup fresh." 14 72
+    --yesno "Problem detected:\n${ip2torWarn}\n\nCheck if service is reachable over Tor:\n${toraddress}/app" 14 72
     if [ "$?" != "1" ]; then
       exit 0
 	  fi