+++ CHANGES.md
@@ -2,7 +2,7 @@
 
 ## What's new in Version 1.7.0 of RaspiBlitz?
 
-- New: Raspberry Pi OS Base Image 64-bit (August 2020)
+- New: Raspberry Pi OS Base Image 64-bit (April 2021)
 - New: Build SD card Image with parameters & FatPack [details](https://github.com/rootzoll/raspiblitz/pull/2044)
 - New: Improve LND uptime and reliability over Tor [details](https://github.com/rootzoll/raspiblitz/pull/2148)
 - New: Lightning Terminal v0.4.1-alpha (Loop, Pool & Faraday UI Bundle) [details](https://github.com/lightninglabs/lightning-terminal#lightning-terminal-lit)
@@ -10,10 +10,11 @@
 - New: Circuitbreaker LND firewall (settings menu) [details](https://github.com/lightningequipment/circuitbreaker/blob/master/README.md)
 - New: Telegraf metrics (experimental) [details](https://github.com/rootzoll/raspiblitz/issues/1369)
 - New: Download whitepaper from blockchain [details](https://github.com/rootzoll/raspiblitz/pull/2017)
-- Update: bitcoin-core version 0.21.0-beta [details](https://github.com/bitcoin/bitcoin/blob/master/doc/release-notes/release-notes-0.21.0.md)
+- New: Extended CONNECT and SYSTEM options in the ssh menu [details](https://github.com/rootzoll/raspiblitz/pull/2119)
+- Update: bitcoin-core version 0.21.0-beta with UPDATE option [details](https://github.com/bitcoin/bitcoin/blob/master/doc/release-notes/release-notes-0.21.0.md)
 - Update: LND version 0.12.1-beta [details](https://github.com/lightningnetwork/lnd/releases/tag/v0.12.1-beta)
 - Update: RTL 0.10.1 [details](https://github.com/Ride-The-Lightning/RTL/releases/tag/v0.10.1)
-- Update: Sphinx-Relay 2.0.3 (always latest release tag & improved connection dialog)
+- Update: Sphinx-Relay 2.0.11 (always latest release tag & improved connection dialog)
 - Update: Thunderhub 0.12.13 [details](https://github.com/apotdevin/thunderhub/releases/tag/v0.12.12)
 - Update: Electrs 0.8.9 [details](https://github.com/romanz/electrs/blob/master/RELEASE-NOTES.md#088-22-feb-2021)
 - Update: BTCPayServer 1.0.7.2 [details](https://github.com/btcpayserver/btcpayserver/releases/tag/v1.0.7.2)